Issue #SB-29145 refactor: Moved Branching Functions To BranchingUtil
